requirements.txt
ansi2html==1.8.0


charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dash==2.9.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.3
fonttools==4.39.0

imageio==2.31.1
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.2
matplotlib==3.7.1
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
packaging==23.0
pandas==1.4.4
Pillow==9.4.0
plotly==5.13.1
pyparsing==3.0.9
python-dateutil==2.8.2

requests==2.28.1
retrying==1.3.4
scikit-image==0.21.0
scikit-learn==1.1.1
scipy==1.10.1

sklearn==0.0
svg.path==6.3
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2023.7.10
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2023.3

Werkzeug==2.2.3
zipp==3.17.
dash_auth==2.3.0
sqlalchemy
redis
dash_bootstrap_components
dash_daq==0.5.0
python-dotenv==1.0.1
# pytz==2022.7.1
# urllib3==1.26.12
# six==1.16.0
# certifi==2022.9.24
#  idna==2.10
#  blinker==1.7.0